float32 output

+++ dafx22_fno/generators/string_solver.py
@@ -169,5 +169,5 @@ class StringSolver():
             y_x[xi,:] = K1_x[xi,:]@ybar
             y_defl_x[xi,:] = cumsum(y_x[xi,:])*T
         
-        return np.real(y_x), np.real(y_defl_x)
+        return np.float32(np.real(y_x)), np.float32(np.real(y_defl_x))
